requirements.txt
asgiref==3.8.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.1.0
Django==4.2.23
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
redis==5.0.4
spotipy==2.23.0
